requirements.txt
ms-swift>=1.6.0
transformers>=4.39
torchmetrics
scikit-learn
cython
pandas
tqdm
numpy
gradio
